pygeoapi/pyproject.toml
[dependency-groups]
geo = [
  "gdal==3.11.2"
]

[project]
name = "pygeoapi-run"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "pygeoapi==0.22.0",
    "shapely>=2.1.2",
    "pyproj>=3.7.2",
    "gunicorn>=23.0.0",
    "gevent>=25.9.1",
    "pyyaml",
    "s3fs",
    "rasterio",
    "django-environ",
    "geopandas>=1.1.1",
    "structlog",
    "typer",
    "schedule>=1.2.2",
]
